ci: ensure that we are always using the latest CUE
Adding the following step to the main CI build for master:

    go get -d cuelang.org/go@latest

in combination with the git clean check, we ensure that cuelang.org is
always referencing the latest CUE, and will fail when it isn't (i.e. we
have just released a new CUE version but not updated cuelang.org). We
could look to automate a PR creation in the future etc, but for now this
is a good fail-safe.