ci: trigger a test workflow on repository_dispatch (#162)
This allows the main CUE repo to trigger a build check on some event in
the main CUE repo, e.g. a new release. This is a basic way to ensure
that cuelang.org is always using the latest CUE version.

Also flip to using GOPROXY=direct for the go get check (so we don't hit
a potentially stale proxy)