ci: check that we are using the latest playground as well as CUE (#167)
This is basically the position we want to be in. Whilst we don't yet
version the playground, we could. At which point we would ensure that we
are using the latest tagged version. For now, this will effectively
default to the tip of playground, which is fine.

Also flip tip to use @main instead of @master for playground.